test imports and name fixes, move 2 tests from ldp to vc
