+++ Cargo.toml
@@ -91,11 +91,14 @@ members = [
 ]
 
 [dev-dependencies]
+async-std = { version = "1.9", features = ["attributes"] }
 blake2_old = { package = "blake2", version = "0.8" } # for bbs doctest
 uuid = { version = "0.8", features = ["v4", "serde"] }
 difference = "2.0"
 tokio = { version = "1.0", features = ["macros"] }
 hyper = { version = "0.14", features = ["server", "http1", "stream"] }
+hex = "0.4"
+multibase = "0.8"
 
 [package.metadata.docs.rs]
 features = ["secp256r1", "secp256k1", "ripemd-160", "http-did"]
