+++ ssi-ldp/Cargo.toml
@@ -31,3 +31,6 @@ chrono = { version = "0.4", features = ["serde"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 chrono = { version = "0.4", features = ["serde", "wasmbind"] }
+
+[dev-dependencies]
+async-std = { version = "1.9", features = ["attributes"] }
