+++ ssi-tzkey/Cargo.toml
@@ -12,7 +12,5 @@ thiserror = "1.0"
 bs58 = { version = "0.4", features = ["check"] }
 
 [dev-dependencies]
-# serde = { version = "1.0", features = ["derive"] }
-# async-std = { version = "1.9", features = ["attributes"] }
 serde_json = "1.0"
 hex = "0.4"
